+++ Samples~/SaveAsyncExample/GameManagerExample.cs
@@ -1,3 +1,5 @@
+// MIT License - Copyright (c) 2025 BUCK Design LLC - https://github.com/buck-co
+
 using UnityEngine;
 using Buck.SaveAsync;